Move most of the long vector test code to its own file
